@@ -1,5 +1,6 @@
 {
   "name": "selfoss-client",
+  "type": "module",
   "dependencies": {
     "@fortawesome/free-brands-svg-icons": "^6.2.0",
     "@fortawesome/free-regular-svg-icons": "^6.2.0",
@@ -11,7 +12,7 @@
     "@szhsin/react-menu": "^4.0.0",
     "classnames": "^2.2.6",
     "core-js": "^3.2.1",
-    "dexie": "^3.0.0",
+    "dexie": "^4.0.0",
     "focus-trap": "^7.0.0",
     "form-urlencoded": "^6.0.0",
     "html-escaper": "^3.0.0",
@@ -21,30 +22,32 @@
     "ramda": "^0.30.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router": "^7.1.1",
     "reset-css": "^5.0.1",
     "rooks": "^7.1.1",
-    "tinykeys": "^2.0.0",
+    "tinykeys": "^3.0.0",
     "unreset-css": "^1.0.1",
     "use-state-with-deps": "^1.1.1",
     "yet-another-react-lightbox": "^3.5.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.0.0",
     "@parcel/packager-raw-url": "^2.0.0",
     "@parcel/transformer-image": "^2.0.1",
     "@parcel/transformer-sass": "^2.0.0",
     "@parcel/transformer-webmanifest": "^2.0.0",
     "autoprefixer": "^10.4.0",
-    "eslint": "^8.45.0",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint": "^9.0.0",
+    "eslint-config-prettier": "^10.0.0",
     "eslint-plugin-react": "^7.20.6",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "globals": "^16.0.0",
     "parcel": "^2.0.0",
     "postcss": "^8.4.4",
     "prettier": "^3.0.0",
     "process": "^0.11.10",
     "stylelint": "^16.0.0",
-    "stylelint-config-standard-scss": "^13.0.0"
+    "stylelint-config-standard-scss": "^14.0.0"
   },
   "scripts": {
     "fix": "npm run fix:js && npm run fix:styles",
@@ -56,7 +59,7 @@
     "fix:js:prettify": "npm run check:js:prettify -- --write",
     "check": "npm run check:js && npm run check:styles",
     "check:js": "npm run check:js:prettify && npm run check:js:lint",
-    "check:js:lint": "eslint --ext .jsx,.js .",
+    "check:js:lint": "eslint",
     "check:js:prettify": "prettier '**.{jsx,js}' --check",
     "check:styles": "npm run check:styles:lint && npm run check:styles:prettify",
     "check:styles:lint": "stylelint styles/*.scss",
