@@ -1,7 +1,7 @@
 import React, { useCallback, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { useLocation, useHistory } from 'react-router-dom';
+import { useLocation, useNavigate } from 'react-router';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { makeEntriesLink } from '../helpers/uri';
 import * as icons from '../icons';
@@ -33,7 +33,7 @@ function joinTerm(words) {
 }
 
 // remove search term
-function handleRemove({ index, location, history, regexSearch }) {
+function handleRemove({ index, location, navigate, regexSearch }) {
     let newterm;
     if (regexSearch) {
         newterm = '';
@@ -46,16 +46,16 @@ function handleRemove({ index, location, history, regexSearch }) {
         newterm = joinTerm(termArray);
     }
 
-    history.push(makeEntriesLink(location, { search: newterm, id: null }));
+    navigate(makeEntriesLink(location, { search: newterm, id: null }));
 }
 
 function SearchWord({ regexSearch, index, item }) {
     const location = useLocation();
-    const history = useHistory();
+    const navigate = useNavigate();
 
     const removeOnClick = useCallback(
-        () => handleRemove({ index, location, history, regexSearch }),
-        [index, location, history, regexSearch],
+        () => handleRemove({ index, location, navigate, regexSearch }),
+        [index, location, navigate, regexSearch],
     );
 
     return (
