@@ -60,7 +60,7 @@ class Index {
             return;
         }
 
-        $this->authentication->needsLoggedInOrPublicMode();
+        $this->authentication->ensureCanRead();
 
         // load tags
         $tags = $this->tagsDao->getWithUnread();
@@ -106,10 +106,32 @@ class Index {
      * @param array<string, mixed> $params request parameters
      * @param array<array{tag: string, color: string, unread: int}> $tags information about tags
      *
-     * @return array{entries: array<array{title: string, strippedTitle: string, content: string, wordCount: int, lengthWithoutTags: int, datetime: string, updatetime: string, sourcetitle: string, tags: StringKeyedArray<array{backColor: string, foreColor: string}>}>, hasMore: bool} html with items
+     * @return array{
+     *     entries: array<array{
+     *         id: int,
+     *         title: string,
+     *         strippedTitle: string,
+     *         content: string,
+     *         unread: bool,
+     *         starred: bool,
+     *         source: int,
+     *         thumbnail: string,
+     *         icon: string,
+     *         uid: string,
+     *         link: string,
+     *         wordCount: int,
+     *         lengthWithoutTags: int,
+     *         datetime: string,
+     *         updatetime: string,
+     *         sourcetitle: string,
+     *         author: string,
+     *         tags: StringKeyedArray<array{backColor: string, foreColor: string}>,
+     *     }>,
+     *     hasMore: bool,
+     * } html with items
      */
     private function loadItems(array $params, array $tags) {
-        $options = ItemOptions::fromUser($params);
+        $options = new ItemOptions($params);
         $entries = [];
         foreach ($this->itemsDao->get($options) as $item) {
             $entries[] = $this->viewHelper->preprocessEntry($item, $this->tagsController, $tags, $options->search);
