@@ -16,6 +16,8 @@ use Monolog\Logger;
 
 /**
  * Helper class for session management.
+ *
+ * This must be the only place to call `session_start()`.
  */
 class Session {
     private bool $started = false;
@@ -60,19 +62,27 @@ class Session {
         }
     }
 
-    public function getBool(string $name, bool $default): bool {
-        if (array_key_exists($name, $_SESSION) && is_bool($_SESSION[$name])) {
+    public function getString(string $name): ?string {
+        $this->start();
+
+        if (array_key_exists($name, $_SESSION) && is_string($_SESSION[$name])) {
             return $_SESSION[$name];
         }
 
-        return $default;
+        return null;
     }
 
-    public function setBool(string $name, bool $value): void {
+    public function setString(string $name, ?string $value): void {
+        $this->start();
+
         $_SESSION[$name] = $value;
     }
 
     public function destroy(): void {
+        if (!$this->started) {
+            return;
+        }
+
         session_destroy();
     }
 }
